+++ deployment/terraform-module-knfsd/main.tf
@@ -20,7 +20,7 @@ terraform {
   }
   provider_meta "aws" {
     user_agent = [
-      "knfsd-file-cache/terraform-module-knfsd/1.1.0-alpha.27"
+      "knfsd-file-cache/terraform-module-knfsd/1.1.0-alpha.28"
     ]
   }
 }