+++ examples/fsx-netapp/main.tf
@@ -23,7 +23,7 @@ terraform {
   }
   provider_meta "aws" {
     user_agent = [
-      "knfsd-file-cache/examples/fsx-netapp/1.1.0-alpha.27"
+      "knfsd-file-cache/examples/fsx-netapp/1.1.0-alpha.28"
     ]
   }
 }