+++ examples/fsx-zfs/main.tf
@@ -11,7 +11,7 @@ terraform {
   }
   provider_meta "aws" {
     user_agent = [
-      "knfsd-file-cache/examples/fsx-zfs/1.1.0-alpha.27"
+      "knfsd-file-cache/examples/fsx-zfs/1.1.0-alpha.28"
     ]
   }
 }